# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DevLens API
  version: 1.0.0
  description: |
    API contract สำหรับ DevLens Engineering Intelligence Platform

    Related documents:
    - 01-business-logic-architecture.md
    - 03-backend-roadmap.md
    - 04-metric-definitions.md
    - 05-database-design.dbml
    - 06-api-design.md
servers:
  - url: /api/v1
    description: Current API version
...
security:
  - bearerAuth: []
paths:
...
  /organizations/{organizationId}/members:
    get:
      tags:
        - Organizations
      summary: Get organization members
      operationId: listOrganizationMembers
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      responses:
        "200":
          description: Organization member list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationMemberListResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags:
        - Organizations
      summary: Add organization member
      operationId: createOrganizationMember
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationMemberRequest"
      responses:
        "201":
          description: Organization member created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationMemberResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /organizations/{organizationId}/members/{memberId}:
    patch:
      tags:
        - Organizations
      summary: Update organization member role
      operationId: updateOrganizationMember
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/MemberId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationMemberRequest"
      responses:
        "200":
          description: Organization member updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationMemberResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - Organizations
      summary: Remove organization member
      operationId: deleteOrganizationMember
      description: This milestone uses hard delete for organization members because `organization_members.deleted_at` is not defined in PostgreSQL schema yet.
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/MemberId"
      responses:
        "204":
          description: Organization member removed
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /organizations/{organizationId}/repositories:
    post:
      tags:
        - Repositories
      summary: Create repository under an organization
      operationId: createRepository
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRepositoryRequest"
      responses:
        "201":
          description: Repository created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    get:
      tags:
        - Repositories
      summary: Get repositories in an organization
      operationId: listRepositories
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/RepositoryStatus"
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/SortBy"
        - $ref: "#/components/parameters/SortOrder"
      responses:
        "200":
          description: Repository list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/ValidationError"
...
  /repositories/{repositoryId}:
    get:
      tags:
        - Repositories
      summary: Get repository detail
      operationId: getRepository
      parameters:
        - $ref: "#/components/parameters/RepositoryId"
      responses:
        "200":
          description: Repository detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    RepositoryId:
      name: repositoryId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    MemberId:
      name: memberId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Search:
      name: search
      in: query
      schema:
        type: string
    SortBy:
      name: sortBy
      in: query
      schema:
        type: string
        example: createdAt
    SortOrder:
      name: sortOrder
      in: query
      schema:
        type: string
        enum:
          - asc
          - desc
        default: desc
    RepositoryStatus:
      name: status
      in: query
      schema:
        type: string
        enum:
          - active
          - inactive
          - archived
...
  responses:
    Unauthorized:
      description: Authentication is required or token is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: User does not have permission to access the resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Request conflicts with current resource state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
  schemas:
...
    PaginationMeta:
      type: object
      required:
        - page
        - pageSize
        - totalItems
        - totalPages
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        totalItems:
          type: integer
          example: 42
        totalPages:
          type: integer
          example: 3
...
    CreateOrganizationMemberRequest:
      type: object
      required:
        - userId
        - role
      properties:
        userId:
          type: string
          format: uuid
        role:
          type: string
          enum:
            - owner
            - admin
            - member
    UpdateOrganizationMemberRequest:
      type: object
      required:
        - role
      properties:
        role:
          type: string
          enum:
            - owner
            - admin
            - member
    OrganizationMember:
      type: object
      required:
        - id
        - organizationId
        - userId
        - role
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        role:
          type: string
          enum:
            - owner
            - admin
            - member
    OrganizationMemberResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/OrganizationMember"
    OrganizationMemberListResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/OrganizationMember"
    Repository:
      type: object
      required:
        - id
        - organizationId
        - githubId
        - name
        - fullName
        - isActive
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        githubId:
          type: integer
          format: int64
        name:
          type: string
          example: devlens-api
        fullName:
          type: string
          example: devlens-labs/devlens-api
        defaultBranch:
          type: string
          example: main
          nullable: true
        isActive:
          type: boolean
        archivedAt:
          type: string
          format: date-time
          nullable: true
        lastSyncedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
...
    CreateRepositoryRequest:
      type: object
      required:
        - githubId
        - name
        - fullName
      properties:
        githubId:
          type: integer
          format: int64
          example: 123456789
        name:
          type: string
          example: devlens-api
        fullName:
          type: string
          example: devlens-labs/devlens-api
        defaultBranch:
          type: string
          example: main
          nullable: true
...
    RepositoryResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/Repository"
    RepositoryListResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Repository"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
...
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: REPOSITORY_NOT_FOUND
        message:
          type: string
          example: Repository not found
        requestId:
          type: string
          example: req_01JABCXYZ
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ValidationIssue:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: from
        message:
          type: string
          example: from must be before or equal to to
    ValidationErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          allOf:
            - $ref: "#/components/schemas/ErrorDetail"
            - type: object
              properties:
                details:
                  type: array
                  items:
                    $ref: "#/components/schemas/ValidationIssue"